:bug:(dockerfile) fix entrypoint
The entrypoint did not work as it could not load environment
variables, so it had to be changed.
We used --use-load-test because the malleus script must have at
least one parameter to work.
The SELENIUM_HUB and SELENIUM_PORT variables were removed because
they were redundant with the .env file.